chore: rename notify_queue to notify
