+++ go.work
@@ -5,5 +5,5 @@ use (
     ./libs/queue
     ./libs/utils
     ./services/greet
-    ./services/notify_queue
+    ./services/notify
 )
