+++ services/notify/main.go
@@ -6,8 +6,9 @@ import (
 	"net/http"
 
 	"github.com/web-notify/api/monorepo/libs/utils/config"
-	consumer "github.com/web-notify/api/monorepo/services/notify_queue/consumer_handler"
-	producer "github.com/web-notify/api/monorepo/services/notify_queue/producer_handler"
+	consumer "github.com/web-notify/api/monorepo/services/notify/consumer_handler"
+	producer "github.com/web-notify/api/monorepo/services/notify/producer_handler"
+	subscribe "github.com/web-notify/api/monorepo/services/notify/subscribe_handler"
 )
 
 func main() {
@@ -15,6 +16,7 @@ func main() {
 	mux := http.NewServeMux()
 	mux.HandleFunc("/api/notify", producer.Handler)
 	mux.HandleFunc("/consumer_handler", consumer.Handler)
+	mux.HandleFunc("/api/subscribe", subscribe.Handler)
 
 	log.Printf("About to listen on %s. Go to https://127.0.0.1:%s/", address, address)
 	log.Fatal(http.ListenAndServe(fmt.Sprintf("127.0.0.1:%s", address), mux))
