+++ services/notify/models/subscription.go
@@ -7,7 +7,6 @@ type Keys struct {
 
 type Subscription struct {
 	Endpoint   string `json:"endpoint"`
-	Message    string `json:"message"`
 	Expiration string `json:"expiration_time"`
 	Keys       Keys   `json:"keys"`
 }
