--- services/notify_queue/go.mod
@@ -1,8 +1,0 @@
-module github.com/web-notify/api/monorepo/services/notify_queue
-
-go 1.18
-
-require (
-	github.com/Azure/azure-pipeline-go v0.1.8 // indirect
-	github.com/Azure/azure-storage-queue-go v0.0.0-20191125232315-636801874cdd // indirect
-)
